Moved styled components out of utils file
