+++ src/js/components/Select/StyledSelect.js
@@ -1,6 +1,14 @@
 import styled from 'styled-components';
 
-import { sizeStyle } from '../../utils';
+import { Button } from '../Button';
+import { DropButton } from '../DropButton';
+import { TextInput } from '../TextInput';
+import {
+  getHoverIndicatorStyle,
+  selectedStyle,
+  controlBorderStyle,
+  sizeStyle,
+} from '../../utils';
 
 export const StyledContainer = styled.div`
   /* IE11 hack to get drop contents to not overflow */
@@ -8,11 +16,50 @@ export const StyledContainer = styled.div`
     width: 100%;
   }
 
-  ${props =>
+  ${(props) =>
     props.dropHeight
       ? sizeStyle('max-height', props.dropHeight, props.theme)
       : 'max-height: inherit;'};
 
-  ${props =>
+  ${(props) =>
     props.theme.select.container && props.theme.select.container.extend};
+`;
+
+// position relative is so scroll can be managed correctly
+export const OptionsBox = styled.div`
+  position: relative;
+  scroll-behavior: smooth;
+  overflow: auto;
+  outline: none;
+`;
+
+export const HiddenInput = styled.input`
+  display: none;
+`;
+
+export const SelectOption = styled(Button)`
+  ${(props) => props.selected && props.textComponent && selectedStyle}
+  // applies theme.global.hover.background to the active
+  // option for mouse and keyboard interactions
+  ${(props) =>
+    props.active &&
+    getHoverIndicatorStyle(
+      !props.children && !props.theme.select.options ? undefined : 'background',
+      props.theme,
+    )}
+  display: block;
+  width: 100%;
+`;
+
+export const SelectTextInput = styled(TextInput)`
+  cursor: ${(props) => (props.defaultCursor ? 'default' : 'pointer')};
+`;
+
+export const StyledSelectDropButton = styled(DropButton)`
+  ${(props) => !props.callerPlain && controlBorderStyle};
+  ${(props) =>
+    props.theme.select &&
+    props.theme.select?.control &&
+    props.theme.select?.control?.extend};
+  ${(props) => props.open && props.theme.select?.control?.open};
 `;
